+++ crates/witgen_macro/src/lib.rs
@@ -1,5 +1,8 @@
 #![deny(warnings)]
 use proc_macro::TokenStream;
+use quote::quote;
+use syn::ItemImpl;
+use witgen_macro_helper::visitor::ImplVisitor;
 
 /// Proc macro attribute to help cargo-witgen to generate right definitions in `.wit` file
 /// ```no_run
@@ -24,5 +27,11 @@ use proc_macro::TokenStream;
 /// ```
 #[proc_macro_attribute]
 pub fn witgen(_attr: TokenStream, item: TokenStream) -> TokenStream {
-    item
+    if let Ok(mut input) = syn::parse::<ItemImpl>(item.clone()) {
+        // This converts attributes paths, e.g. #[path_macro], into a doc string, e.g. ///@path_macro
+        ImplVisitor::path_attrs_to_docs(&mut input);
+        quote! {#input}.into()
+    } else {
+        item
+    }
 }