+++ examples/example_dep/src/lib.rs
@@ -1,0 +1,34 @@
+#![allow(dead_code, clippy::all, clippy::pedantic, clippy::nursery)]
+
+/// Example of an external dependency
+#[witgen::witgen]
+pub type ExternalDep = String;
+
+pub struct SampleResource {}
+
+/// Example Interface
+#[witgen::witgen]
+impl SampleResource {
+  /// Can handle static methods
+  pub fn faa() {}
+ 
+  ///Can add special comments
+  #[payable]
+  pub fn foo(&self) -> String {
+    "foo".to_string()
+  }
+
+  /// Can indicate if mutable
+  pub fn f(&mut self) -> () { () }
+
+}
+
+#[witgen::witgen]
+fn faa() {}
+
+#[witgen::witgen]
+pub enum OtherColors {
+  Orange,
+  Purple,
+  Black,
+}