+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.15.0] - 2022-07-25
+
+### Added
+- function --> func
+- can use `use * from id`; This searches deps for `id`. Currently only top level paths.
+- can use `impl`s, which create's a corresponding `resource` with static methods and instance methods with `@mutable` for methods with `&mut self`.
+- Methods with `#[attribute_macro]` are converted to special doc comments: `///@attribute_macro`
+- Added `aha-wit-parser` version `0.2.0`, to ensure generated wit is valid and will resolve deps used in `use`.
+
 ---
 ## [0.14.0] - 2022-06-14
 ### Added